Fix callback spec to the types actually used.
The last parameter of read/3 is the ID (see the `ProviderModule:read`
call in `erleans_grain:init` and also the implementation in
erleans_provider_ets).

The last parameter of read_by_hash/3 is the hash value (I don't see it
being called, but the implementation seem to expect integer instead of
the grain_ref).